Product Intelligence Tool/backend/backend_requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
pydantic==2.5.3
python-dotenv==1.0.0
apscheduler==3.10.4
httpx==0.26.0
numpy==1.26.3
openai==1.10.0
anthropic==0.18.0
google-analytics-data==0.18.0